Refactoring
- Try to automatically find and validate all paths on startup
- Switching between games now tries to also find and validate all paths
- Some refactoring applied due to different data types in fwg
- Remove hardcoded continents from resource file for Hoi4
- Update submodule

+++ RandomParadox/include/generic/ScenarioGenerator.h
@@ -213,8 +213,8 @@ public:
   // see which country neighbours which
   void evaluateCountryNeighbours();
   virtual void generateCountrySpecifics();
-  void totalResourceVal(const std::vector<double> &resPrev,
-                        double resourceModifier,
+  void totalResourceVal(const std::vector<float> &resPrev,
+                        float resourceModifier,
                         const Scenario::Utils::ResConfig &resourceConfig);
   // calculate how strong each country is
   virtual void evaluateCountries();
